+++ .env
@@ -1,11 +1,13 @@
+# leave "" if you don't want to use the variable.
+
 PROJECT_DIRECTORY="/Users/yuda/Desktop/RIG_v1/project_directory/"               # required
 GPT_MODEL_PATH="/Users/yuda/Desktop/models/gemma-2-2b-it-Q8_0.gguf"             # required
 RAG_MODEL_PATH="/Users/yuda/Desktop/models/bge-m3"                              # required
-RULE_TYPES_DIRECTORY="/Users/yuda/Desktop/RIG_v1/evaluation/data/rule_types"    # optional - you don't nned to change that
-RAG_DIFFERENCE="0.001"                                                          # optional - you don't nned to change that
-RAG_THRESHOLD="0.5"                                                             # optional - you don't nned to change that
-MAX_CONTEXT_LENGTH="1536"                                                       # optional - you don't nned to change that
-MAX_NEW_TOKENS="512"                                                            # optional - you don't nned to change that
-N_THREADS="8"                                                                   # optional - you don't nned to change that
+RULE_TYPES_DIRECTORY="/Users/yuda/Desktop/RIG_v1/evaluation/data/rule_types"    # optional - you don't need to change that
+RAG_DIFFERENCE="0.001"                                                          # optional - you don't need to change that
+RAG_THRESHOLD="0.5"                                                             # optional - you don't need to change that
+MAX_CONTEXT_LENGTH="1536"                                                       # optional - you don't need to change that
+MAX_NEW_TOKENS="512"                                                            # optional - you don't need to change that
+N_THREADS="8"                                                                   # optional - you don't need to change that
 
 LLAMA_SERVER_PATH="/Users/yuda/Desktop/RIG_v1/project_directory/llama.cpp"      # dont need in this version
